+++ forge/src/main/java/de/cech12/woodenhopper/platform/ForgeRegistryHelper.java
@@ -7,6 +7,8 @@ import de.cech12.woodenhopper.blockentity.WoodenHopperBlockEntity;
 import de.cech12.woodenhopper.inventory.WoodenHopperContainer;
 import de.cech12.woodenhopper.platform.services.IRegistryHelper;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.Item;
@@ -22,6 +24,8 @@ import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.Set;
+
 public class ForgeRegistryHelper implements IRegistryHelper {
 
     public static final DeferredRegister<Block> BLOCKS = DeferredRegister.create(ForgeRegistries.BLOCKS, Constants.MOD_ID);
@@ -30,15 +34,15 @@ public class ForgeRegistryHelper implements IRegistryHelper {
     public static final DeferredRegister<MenuType<?>> MENU_TYPES = DeferredRegister.create(ForgeRegistries.Keys.MENU_TYPES, Constants.MOD_ID);
 
     static {
-        RegistryObject<Block> block = BLOCKS.register("wooden_hopper", () -> new WoodenHopperBlock(BlockBehaviour.Properties.of().mapColor(MapColor.WOOD).strength(2.5F).sound(SoundType.WOOD).noOcclusion()));
+        RegistryObject<Block> block = BLOCKS.register(Constants.BLOCK_ITEM_NAME, () -> new WoodenHopperBlock(BlockBehaviour.Properties.of().mapColor(MapColor.WOOD).strength(2.5F).sound(SoundType.WOOD).noOcclusion().setId(ResourceKey.create(BuiltInRegistries.BLOCK.key(), Constants.id(Constants.BLOCK_ITEM_NAME)))));
         Constants.WOODEN_HOPPER_BLOCK = block;
-        Constants.WOODEN_HOPPER_ITEM = fromBlock(block);
-        Constants.WOODEN_HOPPER_BLOCK_ENTITY_TYPE = BLOCK_ENTITY_TYPES.register("wooden_hopper", () -> BlockEntityType.Builder.of(ForgeWoodenHopperBlockEntity::new, Constants.WOODEN_HOPPER_BLOCK.get()).build(null));
+        Constants.WOODEN_HOPPER_ITEM = fromBlock(Constants.BLOCK_ITEM_NAME, block);
+        Constants.WOODEN_HOPPER_BLOCK_ENTITY_TYPE = BLOCK_ENTITY_TYPES.register(Constants.BLOCK_ITEM_NAME, () -> new BlockEntityType<>(ForgeWoodenHopperBlockEntity::new, Set.of(Constants.WOODEN_HOPPER_BLOCK.get())));
         Constants.WOODEN_HOPPER_MENU_TYPE = MENU_TYPES.register("woodenhopper", () -> IForgeMenuType.create((pWindowID, pInventory, pData) -> new WoodenHopperContainer(pWindowID, pInventory)));
     }
 
-    private static RegistryObject<Item> fromBlock(RegistryObject<Block> block) {
-        return ITEMS.register(block.getId().getPath(), () -> new BlockItem(block.get(), new Item.Properties()));
+    private static RegistryObject<Item> fromBlock(String name, RegistryObject<Block> block) {
+        return ITEMS.register(block.getId().getPath(), () -> new BlockItem(block.get(), new Item.Properties().setId(ResourceKey.create(BuiltInRegistries.ITEM.key(), Constants.id(name)))));
     }
 
     @Override