+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Forge Recommended Versioning](https://mcforge.readthedocs.io/en/latest/conventions/versioning/).
 
+## [1.21.3-3.3.0.0] - 2024-11-09
+### Changed
+- updated to Minecraft 1.21.3 (Fabric 0.107.0+1.21.3, Neoforge 21.3.10-beta, Forge 53.0.7)
+- updated Cloth Config support (16.0.141) (Fabric/Quilt)
+- updated ModMenu support (12.0.0-beta.1) (Fabric/Quilt)
+
 ## [1.21.1-3.2.0.0] - 2024-10-15
 - updated to Minecraft 1.21.1 (Fabric 0.105.0+1.21.1, NeoForge 21.1.62, Forge 52.0.21)
 - updated Cloth Config support (15.0.140) (Fabric/Quilt)